0402 修复布局保存代码错误

+++ microservices/query/src/api/models/layouts.py
@@ -1,13 +1,14 @@
 import datetime
-from sqlalchemy import JSON, Column, DateTime, Integer, String
+from sqlalchemy import JSON, Column, DateTime, String, BigInteger, func
 from api.dependencies.database import Base
+
 class SaveRecord(Base):
     __tablename__ = "save_records"
     
     id = Column(String, primary_key=True)  # 使用UUID作为主键
     name = Column(String, nullable=False)  # 保存名称
     version = Column(String, nullable=False)  # 版本号
-    timestamp = Column(Integer, nullable=False)  # 时间戳
+    timestamp = Column(BigInteger, nullable=False)  # 时间戳，注意这里的数据类型已经改为BigInteger
     components = Column(JSON, nullable=False)  # 组件数据
     layouts = Column(JSON, nullable=False)  # 布局数据
     interlinks = Column(JSON, nullable=False)  # 关联数据
